refactor(idmapper): Remove 2 suffix in package name
