+++ pkg/koyeb/apps.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/koyeb/koyeb-api-client-go/api/v1/koyeb"
-	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper2"
+	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper"
 	"github.com/spf13/cobra"
 )
 
@@ -109,13 +109,13 @@ func NewAppHandler() *AppHandler {
 type AppHandler struct {
 	ctxWithAuth context.Context
 	client      *koyeb.APIClient
-	mapper      *idmapper2.Mapper
+	mapper      *idmapper.Mapper
 }
 
 func (h *AppHandler) InitHandler(cmd *cobra.Command, args []string) error {
 	h.client = getApiClient()
 	h.ctxWithAuth = getAuth(context.Background())
-	h.mapper = idmapper2.NewMapper(h.ctxWithAuth, h.client)
+	h.mapper = idmapper.NewMapper(h.ctxWithAuth, h.client)
 	return nil
 }
