+++ pkg/koyeb/apps_describe.go
@@ -2,7 +2,7 @@ package koyeb
 
 import (
 	"github.com/koyeb/koyeb-api-client-go/api/v1/koyeb"
-	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper2"
+	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper"
 	"github.com/koyeb/koyeb-cli/pkg/koyeb/renderer"
 	"github.com/spf13/cobra"
 )
@@ -26,12 +26,12 @@ func (h *AppHandler) Describe(cmd *cobra.Command, args []string) error {
 }
 
 type DescribeAppReply struct {
-	mapper *idmapper2.Mapper
+	mapper *idmapper.Mapper
 	res    *koyeb.GetAppReply
 	full   bool
 }
 
-func NewDescribeAppReply(mapper *idmapper2.Mapper, res *koyeb.GetAppReply, full bool) *DescribeAppReply {
+func NewDescribeAppReply(mapper *idmapper.Mapper, res *koyeb.GetAppReply, full bool) *DescribeAppReply {
 	return &DescribeAppReply{
 		mapper: mapper,
 		res:    res,
