+++ pkg/koyeb/apps_list.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/koyeb/koyeb-api-client-go/api/v1/koyeb"
-	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper2"
+	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper"
 	"github.com/koyeb/koyeb-cli/pkg/koyeb/renderer"
 	"github.com/spf13/cobra"
 )
@@ -39,12 +39,12 @@ func (h *AppHandler) List(cmd *cobra.Command, args []string) error {
 }
 
 type ListAppsReply struct {
-	mapper *idmapper2.Mapper
+	mapper *idmapper.Mapper
 	res    *koyeb.ListAppsReply
 	full   bool
 }
 
-func NewListAppsReply(mapper *idmapper2.Mapper, res *koyeb.ListAppsReply, full bool) *ListAppsReply {
+func NewListAppsReply(mapper *idmapper.Mapper, res *koyeb.ListAppsReply, full bool) *ListAppsReply {
 	return &ListAppsReply{
 		mapper: mapper,
 		res:    res,
