+++ pkg/koyeb/deployments.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/koyeb/koyeb-api-client-go/api/v1/koyeb"
-	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper2"
+	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper"
 	"github.com/spf13/cobra"
 )
 
@@ -61,13 +61,13 @@ func NewDeploymentHandler() *DeploymentHandler {
 type DeploymentHandler struct {
 	ctx    context.Context
 	client *koyeb.APIClient
-	mapper *idmapper2.Mapper
+	mapper *idmapper.Mapper
 }
 
 func (h *DeploymentHandler) InitHandler(cmd *cobra.Command, args []string) error {
 	h.ctx = getAuth(context.Background())
 	h.client = getApiClient()
-	h.mapper = idmapper2.NewMapper(h.ctx, h.client)
+	h.mapper = idmapper.NewMapper(h.ctx, h.client)
 	return nil
 }
