+++ pkg/koyeb/deployments_list.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/koyeb/koyeb-api-client-go/api/v1/koyeb"
-	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper2"
+	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper"
 	"github.com/koyeb/koyeb-cli/pkg/koyeb/renderer"
 	"github.com/spf13/cobra"
 )
@@ -39,12 +39,12 @@ func (h *DeploymentHandler) List(cmd *cobra.Command, args []string) error {
 }
 
 type ListDeploymentsReply struct {
-	mapper *idmapper2.Mapper
+	mapper *idmapper.Mapper
 	res    *koyeb.ListDeploymentsReply
 	full   bool
 }
 
-func NewListDeploymentsReply(mapper *idmapper2.Mapper, res *koyeb.ListDeploymentsReply, full bool) *ListDeploymentsReply {
+func NewListDeploymentsReply(mapper *idmapper.Mapper, res *koyeb.ListDeploymentsReply, full bool) *ListDeploymentsReply {
 	return &ListDeploymentsReply{
 		mapper: mapper,
 		res:    res,
