+++ pkg/koyeb/idmapper/regex.go
@@ -1,4 +1,4 @@
-package idmapper2
+package idmapper
 
 import "regexp"
