+++ pkg/koyeb/instances_get.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/koyeb/koyeb-api-client-go/api/v1/koyeb"
-	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper2"
+	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper"
 	"github.com/koyeb/koyeb-cli/pkg/koyeb/renderer"
 	"github.com/spf13/cobra"
 )
@@ -23,12 +23,12 @@ func (h *InstanceHandler) Get(cmd *cobra.Command, args []string) error {
 }
 
 type GetInstanceReply struct {
-	mapper *idmapper2.Mapper
+	mapper *idmapper.Mapper
 	res    *koyeb.GetInstanceReply
 	full   bool
 }
 
-func NewGetInstanceReply(mapper *idmapper2.Mapper, res *koyeb.GetInstanceReply, full bool) *GetInstanceReply {
+func NewGetInstanceReply(mapper *idmapper.Mapper, res *koyeb.GetInstanceReply, full bool) *GetInstanceReply {
 	return &GetInstanceReply{
 		mapper: mapper,
 		res:    res,
