+++ pkg/koyeb/instances_list.go
@@ -4,7 +4,7 @@ import (
 	"strconv"
 
 	"github.com/koyeb/koyeb-api-client-go/api/v1/koyeb"
-	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper2"
+	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper"
 	"github.com/koyeb/koyeb-cli/pkg/koyeb/renderer"
 	"github.com/spf13/cobra"
 )
@@ -79,12 +79,12 @@ func (h *InstanceHandler) getServiceIDForListQuery(query koyeb.ApiListInstancesR
 }
 
 type ListInstancesReply struct {
-	mapper *idmapper2.Mapper
+	mapper *idmapper.Mapper
 	res    *koyeb.ListInstancesReply
 	full   bool
 }
 
-func NewListInstancesReply(mapper *idmapper2.Mapper, res *koyeb.ListInstancesReply, full bool) *ListInstancesReply {
+func NewListInstancesReply(mapper *idmapper.Mapper, res *koyeb.ListInstancesReply, full bool) *ListInstancesReply {
 	return &ListInstancesReply{
 		mapper: mapper,
 		res:    res,
