+++ pkg/koyeb/secrets_get.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/koyeb/koyeb-api-client-go/api/v1/koyeb"
-	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper2"
+	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper"
 	"github.com/koyeb/koyeb-cli/pkg/koyeb/renderer"
 	"github.com/spf13/cobra"
 )
@@ -24,12 +24,12 @@ func (h *SecretHandler) Get(cmd *cobra.Command, args []string) error {
 }
 
 type GetSecretReply struct {
-	mapper *idmapper2.Mapper
+	mapper *idmapper.Mapper
 	res    *koyeb.GetSecretReply
 	full   bool
 }
 
-func NewGetSecretReply(mapper *idmapper2.Mapper, res *koyeb.GetSecretReply, full bool) *GetSecretReply {
+func NewGetSecretReply(mapper *idmapper.Mapper, res *koyeb.GetSecretReply, full bool) *GetSecretReply {
 	return &GetSecretReply{
 		mapper: mapper,
 		res:    res,
