+++ pkg/koyeb/services_get.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/koyeb/koyeb-api-client-go/api/v1/koyeb"
-	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper2"
+	"github.com/koyeb/koyeb-cli/pkg/koyeb/idmapper"
 	"github.com/koyeb/koyeb-cli/pkg/koyeb/renderer"
 	"github.com/spf13/cobra"
 )
@@ -23,12 +23,12 @@ func (h *ServiceHandler) Get(cmd *cobra.Command, args []string) error {
 }
 
 type GetServiceReply struct {
-	mapper *idmapper2.Mapper
+	mapper *idmapper.Mapper
 	res    *koyeb.GetServiceReply
 	full   bool
 }
 
-func NewGetServiceReply(mapper *idmapper2.Mapper, res *koyeb.GetServiceReply, full bool) *GetServiceReply {
+func NewGetServiceReply(mapper *idmapper.Mapper, res *koyeb.GetServiceReply, full bool) *GetServiceReply {
 	return &GetServiceReply{
 		mapper: mapper,
 		res:    res,
